+++ src/database/db.js
@@ -1,7 +1,7 @@
 import Dexie from 'dexie';
 // Nesse  projeto terei Survey e Form. Cada um terá suas próprias tabelas.
 // Um survey terá várias forms. Cada form estará associada a um survey.
-export const db = new Dexie('db-toformy');
+export const db = new Dexie('db-terrasurvey');
 db.version(1).stores({
   // Crie com created_at e updated_at
   survey: '++id, code, name, created, changed, active',